Make all stateful widgets unit-testable
Flutter's widgets testing framework is nice, but not all third-party
widgets are friendly to be tested by it. Trying to fix third-partys
is not really reasonable, since some PR might never be merged.
Though, there is still a way.
Stateless widgets do not have logic, thus there is no high necessity
in testing them.
Stateful widgets on the other hand always have some sort of logic,
that should be verified, since it might contain bugs in it. Such
logic can be encapsulated in methods of a state, that are easy to
test, and tested with unit-tests.